update changes 02/05/21

+++ basis/src/app/Components/exsisting-customer-details/exsisting-customer-details.component.ts
@@ -6,11 +6,15 @@ import { OrderService } from 'src/app/Services/order.service';
 import { CustomeresService } from 'src/app/Services/customeres.service';
 import { Customer } from 'src/app/Classes/customer';
 import { FormBuilder, FormGroup, Validators } from '@angular/forms';
-import { MatPaginator, MatTableDataSource } from '@angular/material';
+import { MatDialog, MatDialogConfig, MatPaginator, MatTableDataSource } from '@angular/material';
 import { ShoppingBagService } from 'src/app/Services/shopping-bag.service';
 import { ComputerWithProgram, ComputerWithProgramWithDate } from 'src/app/Classes/computer-with-program';
 import { ComputerWithProgramService } from 'src/app/Services/computer-with-program.service';
 import {MatSort} from '@angular/material/sort';
+import Swal from 'sweetalert2';
+import { RulesService } from 'src/app/Services/rules.service';
+import { Software } from 'src/app/Classes/software';
+import { DateAndTimePopupComponent } from '../date-and-time-popup/date-and-time-popup.component';
 
 @Component({
   selector: 'app-exsisting-customer-details',
@@ -25,8 +29,9 @@ export class ExsistingCustomerDetailsComponent implements OnInit , AfterViewInit
   options: FormGroup;
   shoppingBag: Array<ComputerWithProgram>;
   ComputerListWithDate:ComputerWithProgramWithDate[];
-  displayedColumns: string[] = ['Id',"FromDate","EndDate", 'CompanyName', 'Type', 'Prossess','Memory','HardDisk','ScreenSize','Programslist','RemoveItem'];
+  displayedColumns: string[] = ['Id',"FromDate","EndDate", 'CompanyName', 'Type', 'Prossess','Memory','HardDisk','ScreenSize','Programslist','RemoveItem','AddNewOrder'];
   dataSource =null;
+  length=[]
   @ViewChild(MatSort, {static: false}) sort: MatSort;
 
   // @ViewChild(MatPaginator, {static: true}) paginator: MatPaginator;
@@ -35,11 +40,14 @@ export class ExsistingCustomerDetailsComponent implements OnInit , AfterViewInit
   }
 
   constructor(private activatedRoute: ActivatedRoute,
-     private orderSer: OrderService,
+    private orderSer: OrderService,
     private CustomerSer: CustomeresService, 
     private formBuilder: FormBuilder, private router: Router,
     private shoppingBagSer:ShoppingBagService,
-    private compSer:ComputerWithProgramService) { }
+    private compSer:ComputerWithProgramService,
+    private rulesSer:RulesService,
+    private dialog: MatDialog
+    ) { }
     today:Date
   ngOnInit() {
     this.goTo();
@@ -64,7 +72,8 @@ export class ExsistingCustomerDetailsComponent implements OnInit , AfterViewInit
         this.ComputerListWithDate=new Array();
         this.orderList = myData;
         for(let i=0;i<this.orderList.length;i++)
-       { let c=this.compSer.list.find(c=>c.Id=this.orderList[i].ProudactId)
+       {
+          let c=this.compSer.list.find(c=>c.Id==this.orderList[i].ProudactId)
         let tOrF=true
         if(new Date(this.orderList[i].ToDate).getTime()<new Date(this.today).getTime())
         tOrF=true
@@ -96,36 +105,48 @@ export class ExsistingCustomerDetailsComponent implements OnInit , AfterViewInit
     }
     else
       this.router.navigate(['/viewComp']);
-    // this.CustomerSer.GetLandbyCustomerId(userId).subscribe(
-    //   myData => {
-    //     // this.orderList = myData;
-    //   },
-    //   myErr => { console.log(myErr.message); });
-    // console.log(this.orderList);
   }
 
   removeItem(orderID:number)
   {
-   this.orderSer.deleteOrder(orderID).subscribe(myData=>
-    {
-      this.ComputerListWithDate=new Array();
-        this.orderList = myData;
-        for(let i=0;i<this.orderList.length;i++)
-       { let c=this.compSer.list.find(c=>c.Id=this.orderList[i].ProudactId)
-        let tOrF=true
-        if(new Date(this.orderList[i].ToDate).getTime()<new Date(this.today).getTime())
-        tOrF=true
-        else
-        tOrF=false
-       this.ComputerListWithDate.push(new ComputerWithProgramWithDate(c.Id,c.LepTopId,c.CompanyId,c.CompanyName,c.ComputerImg,c.Type,c.Prossess,c.Memory,c.HardDisk,c.ScreenSize,c.ComputerImg,c.Programslist,this.orderList[i].FromDate,this.orderList[i].ToDate,this.orderList[i].Id,tOrF));
-      }
-      this.sortByDueDate(  this.ComputerListWithDate)
-        this.dataSource=new MatTableDataSource<ComputerWithProgramWithDate>(this.ComputerListWithDate);
-        this.dataSource.paginator = this.paginator;
-    
-    })
-  }
+    let o=this.orderList.find(or=>or.Id==orderID)
+    if(new Date(new Date(o.FromDate).setHours(3))<new Date())
+    alert("הזמן עבר...")
+    else
+      Swal.fire({
+        title: 'האם אתה בטוח?',
+        text: 'בעת מחיקת המוצג יגבה סכום ביטול ',
+        cancelButtonColor: "#FF6600",
+        confirmButtonColor: "#FF6600",
+        showCancelButton: true,
+        cancelButtonText: 'Cancle',
+        confirmButtonText: 'Submit',
   
+        allowOutsideClick: () => !Swal.isLoading()
+      })
+        .then((willDelete) => {
+          if (willDelete.isConfirmed) {
+            this.rulesSer.CurrentRull+=20;
+            alert( this.rulesSer.CurrentRull)
+          this.orderSer.deleteOrder(orderID).subscribe(myData=>
+            {
+              this.ComputerListWithDate=new Array();
+                this.orderList = myData;
+                for(let i=0;i<this.orderList.length;i++)
+               { let c=this.compSer.list.find(c=>c.Id=this.orderList[i].ProudactId)
+                let tOrF=true
+                if(new Date(this.orderList[i].ToDate).getTime()<new Date(this.today).getTime())
+                tOrF=true
+                else
+                tOrF=false
+               this.ComputerListWithDate.push(new ComputerWithProgramWithDate(c.Id,c.LepTopId,c.CompanyId,c.CompanyName,c.ComputerImg,c.Type,c.Prossess,c.Memory,c.HardDisk,c.ScreenSize,c.ComputerImg,c.Programslist,this.orderList[i].FromDate,this.orderList[i].ToDate,this.orderList[i].Id,tOrF));
+              }
+              this.sortByDueDate(  this.ComputerListWithDate)
+                this.dataSource=new MatTableDataSource<ComputerWithProgramWithDate>(this.ComputerListWithDate);
+                this.dataSource.paginator = this.paginator;
+            })
+            }});
+          }
 
   ngAfterViewInit() {
     this.dataSource.sort = this.sort;
@@ -175,7 +196,7 @@ export class ExsistingCustomerDetailsComponent implements OnInit , AfterViewInit
         this.sortByDueDate(this.ComputerListWithDate)
             this.dataSource=new MatTableDataSource<ComputerWithProgramWithDate>(this.ComputerListWithDate);
             this.dataSource.paginator = this.paginator;
-         this. cancleChangeDate()
+         this.cancleChangeDate()
           })
       }
     })
@@ -192,6 +213,28 @@ export class ExsistingCustomerDetailsComponent implements OnInit , AfterViewInit
     myArray.sort((a: ComputerWithProgramWithDate, b: ComputerWithProgramWithDate) => {
       return new Date(a.dateE).getTime() - new Date(b.dateE).getTime();
     });
+  }
+
+  applyFilter(event: Event) {
+    const filterValue = (event.target as HTMLInputElement).value;
+    this.dataSource.filter = filterValue.trim().toLowerCase();
+    // this.dataSourceQ.filter = filterValue.trim().valueOf();
+    // this.dataSourceO.filter = filterValue.trim().valueOf();
+  }
+  create(c: ComputerWithProgramWithDate) {
+      const DialogConfig = new MatDialogConfig();
+      DialogConfig.disableClose = true;
+      DialogConfig.autoFocus = true;
+      DialogConfig.width = "60%";
+      DialogConfig.data = { name: new Software(), startDate: String, comp: ComputerWithProgram };
+      DialogConfig.data.startDate = new Date();
+      
+      DialogConfig.data.comp = c;
+      const dialogRef = this.dialog.open(DateAndTimePopupComponent, DialogConfig);
+      dialogRef.afterClosed().subscribe(result => {
+       
+        console.log(result);
+      });
   }
 }
